whole restructering
including functionality tests:
GET
PUT (confirmation & calculation)
PUT (recalculation)
POST (new match)
POST (request JWT token)
