+++ rest/scores.py
@@ -2,9 +2,10 @@
 from flask import request
 from flask_restful import Resource
 from flask_jwt_extended import jwt_required
-from database.models import Score
 from mongoengine import Q
-from ._common import *
+
+from models.score import Score
+from ._common import get_response, handle_error
 
 
 class ScoreApi(Resource):
